backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
google-generativeai
exa-py

